refactor(server): extract checkout status projection
